german_traffic_signs.py contains a more structured, object-oriented, pipeline management system.
classifiers/linear_neural_network.py is a placeholder which will be the home of the current working network in scratchpad_4.signs.py.

test/german_traffic_sign_tests.py will be the home of various tests to ensure integrity of my modeling process. I'd love to learn how to use a unit test framework but no time for that right now.